+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [Parameter.Types.Any] Renamed module to `Parameter.Types.AnyType` to fix elixir warnings
 * [Parameter] Fix `load`, `dump` and `validate` functions to correctly parse Enum values in nested schema.
 * [Parameter.Schema] default options for nested schema is now available.
+* [Parameter.Schema] Add a description on what `use Parameter.Schema` does.
 
 ### Deprecations